Contributte style - added editorconfig - added git attributes - added licence file - added GitHub Actions - added kodiak - ninjify CS
