+++ .gitattributes
@@ -1,0 +1,14 @@
+# Excluded from archive
+.docs export-ignore
+tests export-ignore
+.editorconfig export-ignore
+.gitattributes export-ignore
+.gitignore export-ignore
+.travis.yml export-ignore
+Makefile export-ignore
+phpstan.neon export-ignore
+README.md export-ignore
+ruleset.xml export-ignore
+
+# https://github.com/github/linguist
+*.php* diff=php linguist-language=PHP
