+++ .github/workflows/codesniffer.yml
@@ -1,0 +1,17 @@
+name: "Codesniffer"
+
+on:
+  pull_request:
+
+  push:
+    branches: ["*"]
+
+  schedule:
+    - cron: "0 8 * * 1"
+
+jobs:
+  build:
+    name: "Codesniffer"
+    uses: contributte/.github/.github/workflows/codesniffer.yml@v1
+    with:
+      php: "8.0"
