+++ .github/workflows/phpstan.yml
@@ -1,0 +1,17 @@
+name: "Phpstan"
+
+on:
+  pull_request:
+
+  push:
+    branches: ["*"]
+
+  schedule:
+    - cron: "0 8 * * 1"
+
+jobs:
+  build:
+    name: "Phpstan"
+    uses: contributte/.github/.github/workflows/phpstan.yml@v1
+    with:
+      php: "8.0"
