+++ .github/workflows/tests.yml
@@ -1,0 +1,23 @@
+name: "Nette Tester"
+
+on:
+  pull_request:
+
+  push:
+    branches: ["*"]
+
+  schedule:
+    - cron: "0 8 * * 1"
+
+jobs:
+  build81:
+    name: "Nette Tester"
+    uses: contributte/.github/.github/workflows/nette-tester.yml@v1
+    with:
+      php: "8.1"
+
+  build80:
+    name: "Nette Tester"
+    uses: contributte/.github/.github/workflows/nette-tester.yml@v1
+    with:
+        php: "8.0"
