+++ .gitignore
@@ -1,1 +1,11 @@
-vendor
+# IDE
+.idea
+
+
+# Composer
+/vendor
+/composer.lock
+
+# Tests
+/temp
+/coverage.xml
