+++ composer.json
@@ -28,11 +28,14 @@
         "nette/utils": "^3.2"
     },
     "require-dev": {
-        "nette/tester": "^2.4",
         "mockery/mockery": "^1.4",
-        "phpstan/phpstan": "^1.2",
-        "orisai/coding-standard": "^2.2",
-        "nette/database": "^3.1"
+        "nette/database": "^3.1",
+        "ninjify/nunjuck": "^0.3",
+        "ninjify/qa": "^0.13",
+        "phpstan/phpstan": "^1.0",
+        "phpstan/phpstan-deprecation-rules": "^1.0",
+        "phpstan/phpstan-nette": "^1.0",
+        "phpstan/phpstan-strict-rules": "^1.0"
     },
     "autoload": {
         "psr-4": {
@@ -41,7 +44,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Contributte\\PaginatorControl\\Tests\\": "Tests/src/"
+            "Tests\\Cases\\": "Tests/Cases"
         }
     },
     "minimum-stability": "stable",
