--- phpcs.xml
@@ -1,29 +1,0 @@
-<ruleset
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="./vendor/squizlabs/php_codesniffer/phpcs.xsd">
-
-    <!-- Show only relevant part of path -->
-    <arg name="basepath" value="./"/>
-
-    <!-- Configure cache -->
-    <arg name="cache" value="./Tests/temp/cs/codesniffer.dat"/>
-
-    <!-- Import coding-standard -->
-    <rule ref="./vendor/orisai/coding-standard/src/ruleset-8.0.xml"/>
-
-    <!-- Configure PSR-4 rule -->
-    <rule ref="SlevomatCodingStandard.Files.TypeNameMatchesFileName">
-        <properties>
-            <property name="rootNamespaces" type="array">
-                <element key="src" value="src"/>
-                <element key="Tests" value="Tests"/>
-            </property>
-        </properties>
-    </rule>
-    <rule ref="PSR1.Files.SideEffects.FoundWithSymbols">
-        <exclude-pattern>Tests/*</exclude-pattern>
-    </rule>
-    <rule ref="SlevomatCodingStandard.Files.TypeNameMatchesFileName.NoMatchBetweenTypeNameAndFileName">
-        <exclude-pattern>src/*</exclude-pattern>
-    </rule>
-</ruleset>
