+++ phpstan.neon
@@ -1,9 +1,7 @@
 parameters:
 	scanDirectories:
 		- src
-		- Tests
-	bootstrapFiles:
-		- phpstanConfig.php
+		- tests
 	ignoreErrors:
 		- '#Call to an undefined method Mockery\\ExpectationInterface\|Mockery\\HigherOrderMessage::(with)|(once)|(times)|(withArgs)\(\).#'
 		- '#Access to an undefined property Nette\\Application\\UI\\Templat(e)|(e\|stdClass)::\$(paginator)|(steps).#'
