--- phpstanConfig.php
@@ -1,3 +1,0 @@
-<?php
-
-declare(strict_types=1);
