+++ src/Examples/bootstrap4.latte
@@ -13,5 +13,3 @@
         </li>
     </ul>
 </nav>
-
-
