+++ src/config/common.neon
@@ -1,3 +1,2 @@
 services:
 	- Contributte\PaginatorControl\PaginatorControlFactory
-
