+++ tests/tmp/.gitignore
@@ -1,2 +1,2 @@
-*
 !.gitignore
+*
